Quiz crud works
this is only my work. deleted everything else.
