+++ piJEE/piJEE-web/pom.xml
@@ -42,8 +42,7 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>piJEE-ejb</artifactId>
-            <version>1.0</version>
-            <scope>provided</scope>
+            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>javax</groupId>
