+++ piJEE/piJEE-web/src/main/java/tn/esprit/overpowered/byusforus/managedbeans/QuestionController.java
@@ -1,8 +1,8 @@
-package tn.esprit.overpowered.byusforus.managedbeans.quiz.managedbeans;
+package tn.esprit.overpowered.byusforus.managedbeans;
 
-import tn.esprit.overpowered.byusforus.entities.users.HRManager;
-import tn.esprit.overpowered.byusforus.managedbeans.quiz.managedbeans.util.JsfUtil;
-import tn.esprit.overpowered.byusforus.managedbeans.quiz.managedbeans.util.JsfUtil.PersistAction;
+import tn.esprit.overpowered.byusforus.entities.quiz.Question;
+import tn.esprit.overpowered.byusforus.managedbeans.util.JsfUtil;
+import tn.esprit.overpowered.byusforus.managedbeans.util.JsfUtil.PersistAction;
 
 import java.io.Serializable;
 import java.util.List;
@@ -12,30 +12,31 @@ import java.util.logging.Logger;
 import javax.ejb.EJB;
 import javax.ejb.EJBException;
 import javax.faces.bean.ManagedBean;
-import javax.faces.bean.SessionScoped;
 import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
 import javax.faces.convert.Converter;
 import javax.faces.convert.FacesConverter;
-import tn.esprit.overpowered.byusforus.services.users.HRManagerFacadeLocal;
+import tn.esprit.overpowered.byusforus.services.quiz.QuestionFacadeLocal;
 
 @ManagedBean
-@SessionScoped
-public class HRManagerController implements Serializable {
+@javax.faces.bean.SessionScoped
+public class QuestionController implements Serializable {
+
+    private static final long serialVersionUID = 1L;
 
     @EJB
-    private HRManagerFacadeLocal ejbFacade;
-    private List<HRManager> items = null;
-    private HRManager selected;
+    private QuestionFacadeLocal ejbFacade;
+    private List<Question> items = null;
+    private Question selected;
 
-    public HRManagerController() {
+    public QuestionController() {
     }
 
-    public HRManager getSelected() {
+    public Question getSelected() {
         return selected;
     }
 
-    public void setSelected(HRManager selected) {
+    public void setSelected(Question selected) {
         this.selected = selected;
     }
 
@@ -45,36 +46,36 @@ public class HRManagerController implements Serializable {
     protected void initializeEmbeddableKey() {
     }
 
-    private HRManagerFacadeLocal getFacade() {
+    private QuestionFacadeLocal getFacade() {
         return ejbFacade;
     }
 
-    public HRManager prepareCreate() {
-        selected = new HRManager();
+    public Question prepareCreate() {
+        selected = new Question();
         initializeEmbeddableKey();
         return selected;
     }
 
     public void create() {
-        persist(PersistAction.CREATE, ResourceBundle.getBundle("/Bundle").getString("HRManagerCreated"));
+        persist(PersistAction.CREATE, ResourceBundle.getBundle("/Bundle").getString("QuestionCreated"));
         if (!JsfUtil.isValidationFailed()) {
             items = null;    // Invalidate list of items to trigger re-query.
         }
     }
 
     public void update() {
-        persist(PersistAction.UPDATE, ResourceBundle.getBundle("/Bundle").getString("HRManagerUpdated"));
+        persist(PersistAction.UPDATE, ResourceBundle.getBundle("/Bundle").getString("QuestionUpdated"));
     }
 
     public void destroy() {
-        persist(PersistAction.DELETE, ResourceBundle.getBundle("/Bundle").getString("HRManagerDeleted"));
+        persist(PersistAction.DELETE, ResourceBundle.getBundle("/Bundle").getString("QuestionDeleted"));
         if (!JsfUtil.isValidationFailed()) {
             selected = null; // Remove selection
             items = null;    // Invalidate list of items to trigger re-query.
         }
     }
 
-    public List<HRManager> getItems() {
+    public List<Question> getItems() {
         if (items == null) {
             items = getFacade().findAll();
         }
@@ -109,29 +110,29 @@ public class HRManagerController implements Serializable {
         }
     }
 
-    public HRManager getHRManager(java.lang.Long id) {
+    public Question getQuestion(java.lang.Long id) {
         return getFacade().find(id);
     }
 
-    public List<HRManager> getItemsAvailableSelectMany() {
+    public List<Question> getItemsAvailableSelectMany() {
         return getFacade().findAll();
     }
 
-    public List<HRManager> getItemsAvailableSelectOne() {
+    public List<Question> getItemsAvailableSelectOne() {
         return getFacade().findAll();
     }
 
-    @FacesConverter(forClass = HRManager.class)
-    public static class HRManagerControllerConverter implements Converter {
+    @FacesConverter(forClass = Question.class)
+    public static class QuestionControllerConverter implements Converter {
 
         @Override
         public Object getAsObject(FacesContext facesContext, UIComponent component, String value) {
             if (value == null || value.length() == 0) {
                 return null;
             }
-            HRManagerController controller = (HRManagerController) facesContext.getApplication().getELResolver().
-                    getValue(facesContext.getELContext(), null, "hRManagerController");
-            return controller.getHRManager(getKey(value));
+            QuestionController controller = (QuestionController) facesContext.getApplication().getELResolver().
+                    getValue(facesContext.getELContext(), null, "questionController");
+            return controller.getQuestion(getKey(value));
         }
 
         java.lang.Long getKey(String value) {
@@ -151,11 +152,11 @@ public class HRManagerController implements Serializable {
             if (object == null) {
                 return null;
             }
-            if (object instanceof HRManager) {
-                HRManager o = (HRManager) object;
-                return getStringKey(o.getId());
+            if (object instanceof Question) {
+                Question o = (Question) object;
+                return getStringKey(o.getIdQuestion());
             } else {
-                Logger.getLogger(this.getClass().getName()).log(Level.SEVERE, "object {0} is of type {1}; expected type: {2}", new Object[]{object, object.getClass().getName(), HRManager.class.getName()});
+                Logger.getLogger(this.getClass().getName()).log(Level.SEVERE, "object {0} is of type {1}; expected type: {2}", new Object[]{object, object.getClass().getName(), Question.class.getName()});
                 return null;
             }
         }
