--- piJEE/piJEE-web/src/main/java/tn/esprit/overpowered/byusforus/managedbeans/quiz/managedbeans/CompanyProfileController.java
@@ -1,165 +1,0 @@
-package tn.esprit.overpowered.byusforus.managedbeans.quiz.managedbeans;
-
-import tn.esprit.overpowered.byusforus.entities.users.CompanyProfile;
-import tn.esprit.overpowered.byusforus.managedbeans.quiz.managedbeans.util.JsfUtil;
-import tn.esprit.overpowered.byusforus.managedbeans.quiz.managedbeans.util.JsfUtil.PersistAction;
-
-import java.io.Serializable;
-import java.util.List;
-import java.util.ResourceBundle;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.ejb.EJB;
-import javax.ejb.EJBException;
-import javax.faces.bean.ManagedBean;
-import javax.faces.bean.SessionScoped;
-import javax.faces.component.UIComponent;
-import javax.faces.context.FacesContext;
-import javax.faces.convert.Converter;
-import javax.faces.convert.FacesConverter;
-import tn.esprit.overpowered.byusforus.services.users.CompanyProfileFacadeLocal;
-
-@ManagedBean
-@SessionScoped
-public class CompanyProfileController implements Serializable {
-
-    @EJB
-    private CompanyProfileFacadeLocal ejbFacade;
-    private List<CompanyProfile> items = null;
-    private CompanyProfile selected;
-
-    public CompanyProfileController() {
-    }
-
-    public CompanyProfile getSelected() {
-        return selected;
-    }
-
-    public void setSelected(CompanyProfile selected) {
-        this.selected = selected;
-    }
-
-    protected void setEmbeddableKeys() {
-    }
-
-    protected void initializeEmbeddableKey() {
-    }
-
-    private CompanyProfileFacadeLocal getFacade() {
-        return ejbFacade;
-    }
-
-    public CompanyProfile prepareCreate() {
-        selected = new CompanyProfile();
-        initializeEmbeddableKey();
-        return selected;
-    }
-
-    public void create() {
-        persist(PersistAction.CREATE, ResourceBundle.getBundle("/Bundle").getString("CompanyProfileCreated"));
-        if (!JsfUtil.isValidationFailed()) {
-            items = null;    // Invalidate list of items to trigger re-query.
-        }
-    }
-
-    public void update() {
-        persist(PersistAction.UPDATE, ResourceBundle.getBundle("/Bundle").getString("CompanyProfileUpdated"));
-    }
-
-    public void destroy() {
-        persist(PersistAction.DELETE, ResourceBundle.getBundle("/Bundle").getString("CompanyProfileDeleted"));
-        if (!JsfUtil.isValidationFailed()) {
-            selected = null; // Remove selection
-            items = null;    // Invalidate list of items to trigger re-query.
-        }
-    }
-
-    public List<CompanyProfile> getItems() {
-        if (items == null) {
-            items = getFacade().findAll();
-        }
-        return items;
-    }
-
-    private void persist(PersistAction persistAction, String successMessage) {
-        if (selected != null) {
-            setEmbeddableKeys();
-            try {
-                if (persistAction != PersistAction.DELETE) {
-                    getFacade().edit(selected);
-                } else {
-                    getFacade().remove(selected);
-                }
-                JsfUtil.addSuccessMessage(successMessage);
-            } catch (EJBException ex) {
-                String msg = "";
-                Throwable cause = ex.getCause();
-                if (cause != null) {
-                    msg = cause.getLocalizedMessage();
-                }
-                if (msg.length() > 0) {
-                    JsfUtil.addErrorMessage(msg);
-                } else {
-                    JsfUtil.addErrorMessage(ex, ResourceBundle.getBundle("/Bundle").getString("PersistenceErrorOccured"));
-                }
-            } catch (Exception ex) {
-                Logger.getLogger(this.getClass().getName()).log(Level.SEVERE, null, ex);
-                JsfUtil.addErrorMessage(ex, ResourceBundle.getBundle("/Bundle").getString("PersistenceErrorOccured"));
-            }
-        }
-    }
-
-    public CompanyProfile getCompanyProfile(java.lang.Long id) {
-        return getFacade().find(id);
-    }
-
-    public List<CompanyProfile> getItemsAvailableSelectMany() {
-        return getFacade().findAll();
-    }
-
-    public List<CompanyProfile> getItemsAvailableSelectOne() {
-        return getFacade().findAll();
-    }
-
-    @FacesConverter(forClass = CompanyProfile.class)
-    public static class CompanyProfileControllerConverter implements Converter {
-
-        @Override
-        public Object getAsObject(FacesContext facesContext, UIComponent component, String value) {
-            if (value == null || value.length() == 0) {
-                return null;
-            }
-            CompanyProfileController controller = (CompanyProfileController) facesContext.getApplication().getELResolver().
-                    getValue(facesContext.getELContext(), null, "companyProfileController");
-            return controller.getCompanyProfile(getKey(value));
-        }
-
-        java.lang.Long getKey(String value) {
-            java.lang.Long key;
-            key = Long.valueOf(value);
-            return key;
-        }
-
-        String getStringKey(java.lang.Long value) {
-            StringBuilder sb = new StringBuilder();
-            sb.append(value);
-            return sb.toString();
-        }
-
-        @Override
-        public String getAsString(FacesContext facesContext, UIComponent component, Object object) {
-            if (object == null) {
-                return null;
-            }
-            if (object instanceof CompanyProfile) {
-                CompanyProfile o = (CompanyProfile) object;
-                return getStringKey(o.getId());
-            } else {
-                Logger.getLogger(this.getClass().getName()).log(Level.SEVERE, "object {0} is of type {1}; expected type: {2}", new Object[]{object, object.getClass().getName(), CompanyProfile.class.getName()});
-                return null;
-            }
-        }
-
-    }
-
-}
