+++ piJEE/piJEE-web/src/main/webapp/WEB-INF/faces-config.xml
@@ -9,4 +9,9 @@
             <var>bundle</var>
         </resource-bundle>
     </application>
+    <managed-bean>
+        <managed-bean-name>enums</managed-bean-name>
+        <managed-bean-class>tn.esprit.overpowered.byusforus.managedbeans.util.Enums</managed-bean-class>
+        <managed-bean-scope>application</managed-bean-scope>
+    </managed-bean>
 </faces-config>
