--- piJEE/piJEE-web/src/main/webapp/views/back/companyAdmin/Create.xhtml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="CompanyAdminCreateDlg" widgetVar="CompanyAdminCreateDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.CreateCompanyAdminTitle}">
-            <h:form id="CompanyAdminCreateForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{companyAdminController.selected != null}">
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_companyProfile}" for="companyProfile" />
-                        <p:selectOneMenu id="companyProfile" value="#{companyAdminController.selected.companyProfile}" >
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyProfileItem"
-                                           itemValue="#{companyProfileItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{companyAdminController.selected.skills}" title="#{bundle.CreateCompanyAdminTitle_skills}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{companyAdminController.selected.skills}" title="#{bundle.CreateCompanyAdminTitle_skills}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_company}" for="company" />
-                        <p:selectOneMenu id="company" value="#{companyAdminController.selected.company}" >
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyItem"
-                                           itemValue="#{companyItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_manager}" for="manager" />
-                        <p:selectOneMenu id="manager" value="#{companyAdminController.selected.manager}" >
-                            <f:selectItems value="#{employeeController.itemsAvailableSelectOne}"
-                                           var="managerItem"
-                                           itemValue="#{managerItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_introduction}" for="introduction" />
-                        <p:inputText id="introduction" value="#{companyAdminController.selected.introduction}" title="#{bundle.CreateCompanyAdminTitle_introduction}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_recommendations}" for="recommendations" />
-                        <p:inputText id="recommendations" value="#{companyAdminController.selected.recommendations}" title="#{bundle.CreateCompanyAdminTitle_recommendations}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_activities}" for="activities" />
-                        <p:inputText id="activities" value="#{companyAdminController.selected.activities}" title="#{bundle.CreateCompanyAdminTitle_activities}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_curriculumVitaes}" for="curriculumVitaes" />
-                        <p:inputText id="curriculumVitaes" value="#{companyAdminController.selected.curriculumVitaes}" title="#{bundle.CreateCompanyAdminTitle_curriculumVitaes}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{companyAdminController.selected.skills}" title="#{bundle.CreateCompanyAdminTitle_skills}" required="true" requiredMessage="#{bundle.CreateCompanyAdminRequiredMessage_skills}"/>
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_visits}" for="visits" />
-                        <p:inputText id="visits" value="#{companyAdminController.selected.visits}" title="#{bundle.CreateCompanyAdminTitle_visits}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_recommendedIdList}" for="recommendedIdList" />
-                        <p:inputText id="recommendedIdList" value="#{companyAdminController.selected.recommendedIdList}" title="#{bundle.CreateCompanyAdminTitle_recommendedIdList}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_email}" for="email" />
-                        <p:inputText id="email" value="#{companyAdminController.selected.email}" title="#{bundle.CreateCompanyAdminTitle_email}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_username}" for="username" />
-                        <p:inputText id="username" value="#{companyAdminController.selected.username}" title="#{bundle.CreateCompanyAdminTitle_username}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_firstName}" for="firstName" />
-                        <p:inputText id="firstName" value="#{companyAdminController.selected.firstName}" title="#{bundle.CreateCompanyAdminTitle_firstName}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_lastName}" for="lastName" />
-                        <p:inputText id="lastName" value="#{companyAdminController.selected.lastName}" title="#{bundle.CreateCompanyAdminTitle_lastName}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{companyAdminController.selected.id}" title="#{bundle.CreateCompanyAdminTitle_id}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_salt}" for="salt" />
-                        <p:inputText id="salt" value="#{companyAdminController.selected.salt}" title="#{bundle.CreateCompanyAdminTitle_salt}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyAdminLabel_password}" for="password" />
-                        <p:inputText id="password" value="#{companyAdminController.selected.password}" title="#{bundle.CreateCompanyAdminTitle_password}" />
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{companyAdminController.create}" value="#{bundle.Save}" update="display,:CompanyAdminListForm:datalist,:growl" oncomplete="handleSubmit(args,'CompanyAdminCreateDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="CompanyAdminCreateDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
