--- piJEE/piJEE-web/src/main/webapp/views/back/companyAdmin/Edit.xhtml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="CompanyAdminEditDlg" widgetVar="CompanyAdminEditDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.EditCompanyAdminTitle}">
-            <h:form id="CompanyAdminEditForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{companyAdminController.selected != null}">
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_companyProfile}" for="companyProfile" />
-                        <p:selectOneMenu id="companyProfile" value="#{companyAdminController.selected.companyProfile}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyProfileItem"
-                                           itemValue="#{companyProfileItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{companyAdminController.selected.skills}" title="#{bundle.EditCompanyAdminTitle_skills}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{companyAdminController.selected.skills}" title="#{bundle.EditCompanyAdminTitle_skills}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_company}" for="company" />
-                        <p:selectOneMenu id="company" value="#{companyAdminController.selected.company}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyItem"
-                                           itemValue="#{companyItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_manager}" for="manager" />
-                        <p:selectOneMenu id="manager" value="#{companyAdminController.selected.manager}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{employeeController.itemsAvailableSelectOne}"
-                                           var="managerItem"
-                                           itemValue="#{managerItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_introduction}" for="introduction" />
-                        <p:inputText id="introduction" value="#{companyAdminController.selected.introduction}" title="#{bundle.EditCompanyAdminTitle_introduction}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_recommendations}" for="recommendations" />
-                        <p:inputText id="recommendations" value="#{companyAdminController.selected.recommendations}" title="#{bundle.EditCompanyAdminTitle_recommendations}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_activities}" for="activities" />
-                        <p:inputText id="activities" value="#{companyAdminController.selected.activities}" title="#{bundle.EditCompanyAdminTitle_activities}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_curriculumVitaes}" for="curriculumVitaes" />
-                        <p:inputText id="curriculumVitaes" value="#{companyAdminController.selected.curriculumVitaes}" title="#{bundle.EditCompanyAdminTitle_curriculumVitaes}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{companyAdminController.selected.skills}" title="#{bundle.EditCompanyAdminTitle_skills}" required="true" requiredMessage="#{bundle.EditCompanyAdminRequiredMessage_skills}"/>
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_visits}" for="visits" />
-                        <p:inputText id="visits" value="#{companyAdminController.selected.visits}" title="#{bundle.EditCompanyAdminTitle_visits}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_recommendedIdList}" for="recommendedIdList" />
-                        <p:inputText id="recommendedIdList" value="#{companyAdminController.selected.recommendedIdList}" title="#{bundle.EditCompanyAdminTitle_recommendedIdList}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_email}" for="email" />
-                        <p:inputText id="email" value="#{companyAdminController.selected.email}" title="#{bundle.EditCompanyAdminTitle_email}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_username}" for="username" />
-                        <p:inputText id="username" value="#{companyAdminController.selected.username}" title="#{bundle.EditCompanyAdminTitle_username}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_firstName}" for="firstName" />
-                        <p:inputText id="firstName" value="#{companyAdminController.selected.firstName}" title="#{bundle.EditCompanyAdminTitle_firstName}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_lastName}" for="lastName" />
-                        <p:inputText id="lastName" value="#{companyAdminController.selected.lastName}" title="#{bundle.EditCompanyAdminTitle_lastName}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{companyAdminController.selected.id}" title="#{bundle.EditCompanyAdminTitle_id}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_salt}" for="salt" />
-                        <p:inputText id="salt" value="#{companyAdminController.selected.salt}" title="#{bundle.EditCompanyAdminTitle_salt}" />
-                        <p:outputLabel value="#{bundle.EditCompanyAdminLabel_password}" for="password" />
-                        <p:inputText id="password" value="#{companyAdminController.selected.password}" title="#{bundle.EditCompanyAdminTitle_password}" />
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{companyAdminController.update}" value="#{bundle.Save}" update="display,:CompanyAdminListForm:datalist,:growl" oncomplete="handleSubmit(args, 'CompanyAdminEditDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="CompanyAdminEditDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
