--- piJEE/piJEE-web/src/main/webapp/views/back/companyProfile/Create.xhtml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="CompanyProfileCreateDlg" widgetVar="CompanyProfileCreateDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.CreateCompanyProfileTitle}">
-            <h:form id="CompanyProfileCreateForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{companyProfileController.selected != null}">
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{companyProfileController.selected.id}" title="#{bundle.CreateCompanyProfileTitle_id}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_name}" for="name" />
-                        <p:inputText id="name" value="#{companyProfileController.selected.name}" title="#{bundle.CreateCompanyProfileTitle_name}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_picName}" for="picName" />
-                        <p:inputText id="picName" value="#{companyProfileController.selected.picName}" title="#{bundle.CreateCompanyProfileTitle_picName}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_numViews}" for="numViews" />
-                        <p:inputText id="numViews" value="#{companyProfileController.selected.numViews}" title="#{bundle.CreateCompanyProfileTitle_numViews}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_summary}" for="summary" />
-                        <p:inputText id="summary" value="#{companyProfileController.selected.summary}" title="#{bundle.CreateCompanyProfileTitle_summary}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_sectorOfActivity}" for="sectorOfActivity" />
-                        <p:inputText id="sectorOfActivity" value="#{companyProfileController.selected.sectorOfActivity}" title="#{bundle.CreateCompanyProfileTitle_sectorOfActivity}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_website}" for="website" />
-                        <p:inputText id="website" value="#{companyProfileController.selected.website}" title="#{bundle.CreateCompanyProfileTitle_website}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_companySize}" for="companySize" />
-                        <p:inputText id="companySize" value="#{companyProfileController.selected.companySize}" title="#{bundle.CreateCompanyProfileTitle_companySize}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_dateOfCreation}" for="dateOfCreation" />
-                        <p:inputText id="dateOfCreation" value="#{companyProfileController.selected.dateOfCreation}" title="#{bundle.CreateCompanyProfileTitle_dateOfCreation}" />
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_companyAdmin}" for="companyAdmin" />
-                        <p:selectOneMenu id="companyAdmin" value="#{companyProfileController.selected.companyAdmin}" >
-                            <f:selectItems value="#{companyAdminController.itemsAvailableSelectOne}"
-                                           var="companyAdminItem"
-                                           itemValue="#{companyAdminItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateCompanyProfileLabel_companyHRManager}" for="companyHRManager" />
-                        <p:selectOneMenu id="companyHRManager" value="#{companyProfileController.selected.companyHRManager}" >
-                            <f:selectItems value="#{hRManagerController.itemsAvailableSelectOne}"
-                                           var="companyHRManagerItem"
-                                           itemValue="#{companyHRManagerItem}"/>
-                        </p:selectOneMenu>
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{companyProfileController.create}" value="#{bundle.Save}" update="display,:CompanyProfileListForm:datalist,:growl" oncomplete="handleSubmit(args,'CompanyProfileCreateDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="CompanyProfileCreateDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
