--- piJEE/piJEE-web/src/main/webapp/views/back/companyProfile/View.xhtml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="CompanyProfileViewDlg" widgetVar="CompanyProfileViewDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.ViewCompanyProfileTitle}">
-            <h:form id="CompanyProfileViewForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{companyProfileController.selected != null}">
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_id}"/>
-                        <h:outputText value="#{companyProfileController.selected.id}" title="#{bundle.ViewCompanyProfileTitle_id}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_name}"/>
-                        <h:outputText value="#{companyProfileController.selected.name}" title="#{bundle.ViewCompanyProfileTitle_name}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_picName}"/>
-                        <h:outputText value="#{companyProfileController.selected.picName}" title="#{bundle.ViewCompanyProfileTitle_picName}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_numViews}"/>
-                        <h:outputText value="#{companyProfileController.selected.numViews}" title="#{bundle.ViewCompanyProfileTitle_numViews}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_summary}"/>
-                        <h:outputText value="#{companyProfileController.selected.summary}" title="#{bundle.ViewCompanyProfileTitle_summary}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_sectorOfActivity}"/>
-                        <h:outputText value="#{companyProfileController.selected.sectorOfActivity}" title="#{bundle.ViewCompanyProfileTitle_sectorOfActivity}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_website}"/>
-                        <h:outputText value="#{companyProfileController.selected.website}" title="#{bundle.ViewCompanyProfileTitle_website}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_companySize}"/>
-                        <h:outputText value="#{companyProfileController.selected.companySize}" title="#{bundle.ViewCompanyProfileTitle_companySize}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_dateOfCreation}"/>
-                        <h:outputText value="#{companyProfileController.selected.dateOfCreation}" title="#{bundle.ViewCompanyProfileTitle_dateOfCreation}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_companyAdmin}"/>
-                        <h:outputText value="#{companyProfileController.selected.companyAdmin}" title="#{bundle.ViewCompanyProfileTitle_companyAdmin}"/>
-                        <h:outputText value="#{bundle.ViewCompanyProfileLabel_companyHRManager}"/>
-                        <h:outputText value="#{companyProfileController.selected.companyHRManager}" title="#{bundle.ViewCompanyProfileTitle_companyHRManager}"/>
-                    </p:panelGrid>
-                    <p:commandButton value="#{bundle.Close}" onclick="CompanyProfileViewDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
