--- piJEE/piJEE-web/src/main/webapp/views/back/employee/Create.xhtml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="EmployeeCreateDlg" widgetVar="EmployeeCreateDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.CreateEmployeeTitle}">
-            <h:form id="EmployeeCreateForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{employeeController.selected != null}">
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{employeeController.selected.skills}" title="#{bundle.CreateEmployeeTitle_skills}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_company}" for="company" />
-                        <p:selectOneMenu id="company" value="#{employeeController.selected.company}" >
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyItem"
-                                           itemValue="#{companyItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_manager}" for="manager" />
-                        <p:selectOneMenu id="manager" value="#{employeeController.selected.manager}" >
-                            <f:selectItems value="#{employeeController.itemsAvailableSelectOne}"
-                                           var="managerItem"
-                                           itemValue="#{managerItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_introduction}" for="introduction" />
-                        <p:inputText id="introduction" value="#{employeeController.selected.introduction}" title="#{bundle.CreateEmployeeTitle_introduction}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_recommendations}" for="recommendations" />
-                        <p:inputText id="recommendations" value="#{employeeController.selected.recommendations}" title="#{bundle.CreateEmployeeTitle_recommendations}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_activities}" for="activities" />
-                        <p:inputText id="activities" value="#{employeeController.selected.activities}" title="#{bundle.CreateEmployeeTitle_activities}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_curriculumVitaes}" for="curriculumVitaes" />
-                        <p:inputText id="curriculumVitaes" value="#{employeeController.selected.curriculumVitaes}" title="#{bundle.CreateEmployeeTitle_curriculumVitaes}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{employeeController.selected.skills}" title="#{bundle.CreateEmployeeTitle_skills}" required="true" requiredMessage="#{bundle.CreateEmployeeRequiredMessage_skills}"/>
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_visits}" for="visits" />
-                        <p:inputText id="visits" value="#{employeeController.selected.visits}" title="#{bundle.CreateEmployeeTitle_visits}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_recommendedIdList}" for="recommendedIdList" />
-                        <p:inputText id="recommendedIdList" value="#{employeeController.selected.recommendedIdList}" title="#{bundle.CreateEmployeeTitle_recommendedIdList}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_email}" for="email" />
-                        <p:inputText id="email" value="#{employeeController.selected.email}" title="#{bundle.CreateEmployeeTitle_email}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_username}" for="username" />
-                        <p:inputText id="username" value="#{employeeController.selected.username}" title="#{bundle.CreateEmployeeTitle_username}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_firstName}" for="firstName" />
-                        <p:inputText id="firstName" value="#{employeeController.selected.firstName}" title="#{bundle.CreateEmployeeTitle_firstName}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_lastName}" for="lastName" />
-                        <p:inputText id="lastName" value="#{employeeController.selected.lastName}" title="#{bundle.CreateEmployeeTitle_lastName}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{employeeController.selected.id}" title="#{bundle.CreateEmployeeTitle_id}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_salt}" for="salt" />
-                        <p:inputText id="salt" value="#{employeeController.selected.salt}" title="#{bundle.CreateEmployeeTitle_salt}" />
-                        <p:outputLabel value="#{bundle.CreateEmployeeLabel_password}" for="password" />
-                        <p:inputText id="password" value="#{employeeController.selected.password}" title="#{bundle.CreateEmployeeTitle_password}" />
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{employeeController.create}" value="#{bundle.Save}" update="display,:EmployeeListForm:datalist,:growl" oncomplete="handleSubmit(args,'EmployeeCreateDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="EmployeeCreateDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
