--- piJEE/piJEE-web/src/main/webapp/views/back/employee/Edit.xhtml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="EmployeeEditDlg" widgetVar="EmployeeEditDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.EditEmployeeTitle}">
-            <h:form id="EmployeeEditForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{employeeController.selected != null}">
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{employeeController.selected.skills}" title="#{bundle.EditEmployeeTitle_skills}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_company}" for="company" />
-                        <p:selectOneMenu id="company" value="#{employeeController.selected.company}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyItem"
-                                           itemValue="#{companyItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_manager}" for="manager" />
-                        <p:selectOneMenu id="manager" value="#{employeeController.selected.manager}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{employeeController.itemsAvailableSelectOne}"
-                                           var="managerItem"
-                                           itemValue="#{managerItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_introduction}" for="introduction" />
-                        <p:inputText id="introduction" value="#{employeeController.selected.introduction}" title="#{bundle.EditEmployeeTitle_introduction}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_recommendations}" for="recommendations" />
-                        <p:inputText id="recommendations" value="#{employeeController.selected.recommendations}" title="#{bundle.EditEmployeeTitle_recommendations}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_activities}" for="activities" />
-                        <p:inputText id="activities" value="#{employeeController.selected.activities}" title="#{bundle.EditEmployeeTitle_activities}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_curriculumVitaes}" for="curriculumVitaes" />
-                        <p:inputText id="curriculumVitaes" value="#{employeeController.selected.curriculumVitaes}" title="#{bundle.EditEmployeeTitle_curriculumVitaes}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{employeeController.selected.skills}" title="#{bundle.EditEmployeeTitle_skills}" required="true" requiredMessage="#{bundle.EditEmployeeRequiredMessage_skills}"/>
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_visits}" for="visits" />
-                        <p:inputText id="visits" value="#{employeeController.selected.visits}" title="#{bundle.EditEmployeeTitle_visits}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_recommendedIdList}" for="recommendedIdList" />
-                        <p:inputText id="recommendedIdList" value="#{employeeController.selected.recommendedIdList}" title="#{bundle.EditEmployeeTitle_recommendedIdList}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_email}" for="email" />
-                        <p:inputText id="email" value="#{employeeController.selected.email}" title="#{bundle.EditEmployeeTitle_email}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_username}" for="username" />
-                        <p:inputText id="username" value="#{employeeController.selected.username}" title="#{bundle.EditEmployeeTitle_username}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_firstName}" for="firstName" />
-                        <p:inputText id="firstName" value="#{employeeController.selected.firstName}" title="#{bundle.EditEmployeeTitle_firstName}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_lastName}" for="lastName" />
-                        <p:inputText id="lastName" value="#{employeeController.selected.lastName}" title="#{bundle.EditEmployeeTitle_lastName}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{employeeController.selected.id}" title="#{bundle.EditEmployeeTitle_id}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_salt}" for="salt" />
-                        <p:inputText id="salt" value="#{employeeController.selected.salt}" title="#{bundle.EditEmployeeTitle_salt}" />
-                        <p:outputLabel value="#{bundle.EditEmployeeLabel_password}" for="password" />
-                        <p:inputText id="password" value="#{employeeController.selected.password}" title="#{bundle.EditEmployeeTitle_password}" />
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{employeeController.update}" value="#{bundle.Save}" update="display,:EmployeeListForm:datalist,:growl" oncomplete="handleSubmit(args, 'EmployeeEditDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="EmployeeEditDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
