--- piJEE/piJEE-web/src/main/webapp/views/back/employee/View.xhtml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="EmployeeViewDlg" widgetVar="EmployeeViewDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.ViewEmployeeTitle}">
-            <h:form id="EmployeeViewForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{employeeController.selected != null}">
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_skills}"/>
-                        <h:outputText value="#{employeeController.selected.skills}" title="#{bundle.ViewEmployeeTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_company}"/>
-                        <h:outputText value="#{employeeController.selected.company}" title="#{bundle.ViewEmployeeTitle_company}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_manager}"/>
-                        <h:outputText value="#{employeeController.selected.manager}" title="#{bundle.ViewEmployeeTitle_manager}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_introduction}"/>
-                        <h:outputText value="#{employeeController.selected.introduction}" title="#{bundle.ViewEmployeeTitle_introduction}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_recommendations}"/>
-                        <h:outputText value="#{employeeController.selected.recommendations}" title="#{bundle.ViewEmployeeTitle_recommendations}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_activities}"/>
-                        <h:outputText value="#{employeeController.selected.activities}" title="#{bundle.ViewEmployeeTitle_activities}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_curriculumVitaes}"/>
-                        <h:outputText value="#{employeeController.selected.curriculumVitaes}" title="#{bundle.ViewEmployeeTitle_curriculumVitaes}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_skills}"/>
-                        <h:outputText value="#{employeeController.selected.skills}" title="#{bundle.ViewEmployeeTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_visits}"/>
-                        <h:outputText value="#{employeeController.selected.visits}" title="#{bundle.ViewEmployeeTitle_visits}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_recommendedIdList}"/>
-                        <h:outputText value="#{employeeController.selected.recommendedIdList}" title="#{bundle.ViewEmployeeTitle_recommendedIdList}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_email}"/>
-                        <h:outputText value="#{employeeController.selected.email}" title="#{bundle.ViewEmployeeTitle_email}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_username}"/>
-                        <h:outputText value="#{employeeController.selected.username}" title="#{bundle.ViewEmployeeTitle_username}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_firstName}"/>
-                        <h:outputText value="#{employeeController.selected.firstName}" title="#{bundle.ViewEmployeeTitle_firstName}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_lastName}"/>
-                        <h:outputText value="#{employeeController.selected.lastName}" title="#{bundle.ViewEmployeeTitle_lastName}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_id}"/>
-                        <h:outputText value="#{employeeController.selected.id}" title="#{bundle.ViewEmployeeTitle_id}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_salt}"/>
-                        <h:outputText value="#{employeeController.selected.salt}" title="#{bundle.ViewEmployeeTitle_salt}"/>
-                        <h:outputText value="#{bundle.ViewEmployeeLabel_password}"/>
-                        <h:outputText value="#{employeeController.selected.password}" title="#{bundle.ViewEmployeeTitle_password}"/>
-                    </p:panelGrid>
-                    <p:commandButton value="#{bundle.Close}" onclick="EmployeeViewDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
