--- piJEE/piJEE-web/src/main/webapp/views/back/hRManager/Edit.xhtml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="HRManagerEditDlg" widgetVar="HRManagerEditDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.EditHRManagerTitle}">
-            <h:form id="HRManagerEditForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{hRManagerController.selected != null}">
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_companyProfile}" for="companyProfile" />
-                        <p:selectOneMenu id="companyProfile" value="#{hRManagerController.selected.companyProfile}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyProfileItem"
-                                           itemValue="#{companyProfileItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{hRManagerController.selected.skills}" title="#{bundle.EditHRManagerTitle_skills}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_company}" for="company" />
-                        <p:inputText id="company" value="#{hRManagerController.selected.company}" title="#{bundle.EditHRManagerTitle_company}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{hRManagerController.selected.skills}" title="#{bundle.EditHRManagerTitle_skills}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_company}" for="company" />
-                        <p:selectOneMenu id="company" value="#{hRManagerController.selected.company}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyItem"
-                                           itemValue="#{companyItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_manager}" for="manager" />
-                        <p:selectOneMenu id="manager" value="#{hRManagerController.selected.manager}" >
-                            <f:selectItem itemLabel="#{bundle.SelectOneMessage}"/>
-                            <f:selectItems value="#{employeeController.itemsAvailableSelectOne}"
-                                           var="managerItem"
-                                           itemValue="#{managerItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_introduction}" for="introduction" />
-                        <p:inputText id="introduction" value="#{hRManagerController.selected.introduction}" title="#{bundle.EditHRManagerTitle_introduction}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_recommendations}" for="recommendations" />
-                        <p:inputText id="recommendations" value="#{hRManagerController.selected.recommendations}" title="#{bundle.EditHRManagerTitle_recommendations}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_activities}" for="activities" />
-                        <p:inputText id="activities" value="#{hRManagerController.selected.activities}" title="#{bundle.EditHRManagerTitle_activities}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_curriculumVitaes}" for="curriculumVitaes" />
-                        <p:inputText id="curriculumVitaes" value="#{hRManagerController.selected.curriculumVitaes}" title="#{bundle.EditHRManagerTitle_curriculumVitaes}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{hRManagerController.selected.skills}" title="#{bundle.EditHRManagerTitle_skills}" required="true" requiredMessage="#{bundle.EditHRManagerRequiredMessage_skills}"/>
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_visits}" for="visits" />
-                        <p:inputText id="visits" value="#{hRManagerController.selected.visits}" title="#{bundle.EditHRManagerTitle_visits}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_recommendedIdList}" for="recommendedIdList" />
-                        <p:inputText id="recommendedIdList" value="#{hRManagerController.selected.recommendedIdList}" title="#{bundle.EditHRManagerTitle_recommendedIdList}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_email}" for="email" />
-                        <p:inputText id="email" value="#{hRManagerController.selected.email}" title="#{bundle.EditHRManagerTitle_email}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_username}" for="username" />
-                        <p:inputText id="username" value="#{hRManagerController.selected.username}" title="#{bundle.EditHRManagerTitle_username}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_firstName}" for="firstName" />
-                        <p:inputText id="firstName" value="#{hRManagerController.selected.firstName}" title="#{bundle.EditHRManagerTitle_firstName}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_lastName}" for="lastName" />
-                        <p:inputText id="lastName" value="#{hRManagerController.selected.lastName}" title="#{bundle.EditHRManagerTitle_lastName}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{hRManagerController.selected.id}" title="#{bundle.EditHRManagerTitle_id}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_salt}" for="salt" />
-                        <p:inputText id="salt" value="#{hRManagerController.selected.salt}" title="#{bundle.EditHRManagerTitle_salt}" />
-                        <p:outputLabel value="#{bundle.EditHRManagerLabel_password}" for="password" />
-                        <p:inputText id="password" value="#{hRManagerController.selected.password}" title="#{bundle.EditHRManagerTitle_password}" />
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{hRManagerController.update}" value="#{bundle.Save}" update="display,:HRManagerListForm:datalist,:growl" oncomplete="handleSubmit(args, 'HRManagerEditDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="HRManagerEditDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
