--- piJEE/piJEE-web/src/main/webapp/views/back/hRManager/View.xhtml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="HRManagerViewDlg" widgetVar="HRManagerViewDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.ViewHRManagerTitle}">
-            <h:form id="HRManagerViewForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{hRManagerController.selected != null}">
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_companyProfile}"/>
-                        <h:outputText value="#{hRManagerController.selected.companyProfile}" title="#{bundle.ViewHRManagerTitle_companyProfile}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_skills}"/>
-                        <h:outputText value="#{hRManagerController.selected.skills}" title="#{bundle.ViewHRManagerTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_company}"/>
-                        <h:outputText value="#{hRManagerController.selected.company}" title="#{bundle.ViewHRManagerTitle_company}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_skills}"/>
-                        <h:outputText value="#{hRManagerController.selected.skills}" title="#{bundle.ViewHRManagerTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_company}"/>
-                        <h:outputText value="#{hRManagerController.selected.company}" title="#{bundle.ViewHRManagerTitle_company}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_manager}"/>
-                        <h:outputText value="#{hRManagerController.selected.manager}" title="#{bundle.ViewHRManagerTitle_manager}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_introduction}"/>
-                        <h:outputText value="#{hRManagerController.selected.introduction}" title="#{bundle.ViewHRManagerTitle_introduction}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_recommendations}"/>
-                        <h:outputText value="#{hRManagerController.selected.recommendations}" title="#{bundle.ViewHRManagerTitle_recommendations}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_activities}"/>
-                        <h:outputText value="#{hRManagerController.selected.activities}" title="#{bundle.ViewHRManagerTitle_activities}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_curriculumVitaes}"/>
-                        <h:outputText value="#{hRManagerController.selected.curriculumVitaes}" title="#{bundle.ViewHRManagerTitle_curriculumVitaes}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_skills}"/>
-                        <h:outputText value="#{hRManagerController.selected.skills}" title="#{bundle.ViewHRManagerTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_visits}"/>
-                        <h:outputText value="#{hRManagerController.selected.visits}" title="#{bundle.ViewHRManagerTitle_visits}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_recommendedIdList}"/>
-                        <h:outputText value="#{hRManagerController.selected.recommendedIdList}" title="#{bundle.ViewHRManagerTitle_recommendedIdList}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_email}"/>
-                        <h:outputText value="#{hRManagerController.selected.email}" title="#{bundle.ViewHRManagerTitle_email}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_username}"/>
-                        <h:outputText value="#{hRManagerController.selected.username}" title="#{bundle.ViewHRManagerTitle_username}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_firstName}"/>
-                        <h:outputText value="#{hRManagerController.selected.firstName}" title="#{bundle.ViewHRManagerTitle_firstName}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_lastName}"/>
-                        <h:outputText value="#{hRManagerController.selected.lastName}" title="#{bundle.ViewHRManagerTitle_lastName}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_id}"/>
-                        <h:outputText value="#{hRManagerController.selected.id}" title="#{bundle.ViewHRManagerTitle_id}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_salt}"/>
-                        <h:outputText value="#{hRManagerController.selected.salt}" title="#{bundle.ViewHRManagerTitle_salt}"/>
-                        <h:outputText value="#{bundle.ViewHRManagerLabel_password}"/>
-                        <h:outputText value="#{hRManagerController.selected.password}" title="#{bundle.ViewHRManagerTitle_password}"/>
-                    </p:panelGrid>
-                    <p:commandButton value="#{bundle.Close}" onclick="HRManagerViewDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
