--- piJEE/piJEE-web/src/main/webapp/views/back/jobOffer/List.xhtml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition template="/template.xhtml">
-
-
-        <ui:define name="title">
-            <h:outputText value="#{bundle.ListJobOfferTitle}"></h:outputText>
-        </ui:define>
-
-        <ui:define name="body">
-            <h:form id="JobOfferListForm">
-                <p:panel header="#{bundle.ListJobOfferTitle}">
-                    <p:dataTable id="datalist" value="#{jobOfferController.items}" var="item"
-                                 selectionMode="single" selection="#{jobOfferController.selected}"
-                                 paginator="true"
-                                 rowKey="#{item.id}"
-                                 rows="10"
-                                 rowsPerPageTemplate="10,20,30,40,50"
-                                 >
-
-                        <p:ajax event="rowSelect"   update="createButton viewButton editButton deleteButton"/>
-                        <p:ajax event="rowUnselect" update="createButton viewButton editButton deleteButton"/>
-
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_id}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.id}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_title}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.title}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_dateOfCreation}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.dateOfCreation}">
-                                <f:convertDateTime pattern="MM/dd/yyyy" />
-                            </h:outputText>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_description}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.description}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_city}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.city}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_offerStatus}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.offerStatus}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_dateOfArchive}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.dateOfArchive}">
-                                <f:convertDateTime pattern="MM/dd/yyyy" />
-                            </h:outputText>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_expertiseLevel}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.expertiseLevel}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_skills}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.skills}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_peopleNeeded}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.peopleNeeded}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_company}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.company}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_hRManager}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.hRManager}"/>
-                        </p:column>
-                        <p:column>
-                            <f:facet name="header">
-                                <h:outputText value="#{bundle.ListJobOfferTitle_pManager}"/>
-                            </f:facet>
-                            <h:outputText value="#{item.pManager}"/>
-                        </p:column>
-                        <f:facet name="footer">
-                            <p:commandButton id="createButton" icon="ui-icon-plus"   value="#{bundle.Create}" actionListener="#{jobOfferController.prepareCreate}" update=":JobOfferCreateForm" oncomplete="PF('JobOfferCreateDialog').show()"/>
-                            <p:commandButton id="viewButton"   icon="ui-icon-search" value="#{bundle.View}" update=":JobOfferViewForm" oncomplete="PF('JobOfferViewDialog').show()" disabled="#{empty jobOfferController.selected}"/>
-                            <p:commandButton id="editButton"   icon="ui-icon-pencil" value="#{bundle.Edit}" update=":JobOfferEditForm" oncomplete="PF('JobOfferEditDialog').show()" disabled="#{empty jobOfferController.selected}"/>
-                            <p:commandButton id="deleteButton" icon="ui-icon-trash"  value="#{bundle.Delete}" actionListener="#{jobOfferController.destroy}" update=":growl,datalist" disabled="#{empty jobOfferController.selected}"/>
-                        </f:facet>
-                    </p:dataTable>
-                </p:panel>
-            </h:form>
-
-            <ui:include src="Create.xhtml"/>
-            <ui:include src="Edit.xhtml"/>
-            <ui:include src="View.xhtml"/>
-        </ui:define>
-    </ui:composition>
-
-</html>
