--- piJEE/piJEE-web/src/main/webapp/views/back/projectManager/Create.xhtml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="ProjectManagerCreateDlg" widgetVar="ProjectManagerCreateDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.CreateProjectManagerTitle}">
-            <h:form id="ProjectManagerCreateForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{projectManagerController.selected != null}">
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_companyProfile}" for="companyProfile" />
-                        <p:selectOneMenu id="companyProfile" value="#{projectManagerController.selected.companyProfile}" >
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyProfileItem"
-                                           itemValue="#{companyProfileItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{projectManagerController.selected.skills}" title="#{bundle.CreateProjectManagerTitle_skills}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{projectManagerController.selected.skills}" title="#{bundle.CreateProjectManagerTitle_skills}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_company}" for="company" />
-                        <p:selectOneMenu id="company" value="#{projectManagerController.selected.company}" >
-                            <f:selectItems value="#{companyProfileController.itemsAvailableSelectOne}"
-                                           var="companyItem"
-                                           itemValue="#{companyItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_manager}" for="manager" />
-                        <p:selectOneMenu id="manager" value="#{projectManagerController.selected.manager}" >
-                            <f:selectItems value="#{employeeController.itemsAvailableSelectOne}"
-                                           var="managerItem"
-                                           itemValue="#{managerItem}"/>
-                        </p:selectOneMenu>
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_introduction}" for="introduction" />
-                        <p:inputText id="introduction" value="#{projectManagerController.selected.introduction}" title="#{bundle.CreateProjectManagerTitle_introduction}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_recommendations}" for="recommendations" />
-                        <p:inputText id="recommendations" value="#{projectManagerController.selected.recommendations}" title="#{bundle.CreateProjectManagerTitle_recommendations}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_activities}" for="activities" />
-                        <p:inputText id="activities" value="#{projectManagerController.selected.activities}" title="#{bundle.CreateProjectManagerTitle_activities}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_curriculumVitaes}" for="curriculumVitaes" />
-                        <p:inputText id="curriculumVitaes" value="#{projectManagerController.selected.curriculumVitaes}" title="#{bundle.CreateProjectManagerTitle_curriculumVitaes}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_skills}" for="skills" />
-                        <p:inputText id="skills" value="#{projectManagerController.selected.skills}" title="#{bundle.CreateProjectManagerTitle_skills}" required="true" requiredMessage="#{bundle.CreateProjectManagerRequiredMessage_skills}"/>
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_visits}" for="visits" />
-                        <p:inputText id="visits" value="#{projectManagerController.selected.visits}" title="#{bundle.CreateProjectManagerTitle_visits}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_recommendedIdList}" for="recommendedIdList" />
-                        <p:inputText id="recommendedIdList" value="#{projectManagerController.selected.recommendedIdList}" title="#{bundle.CreateProjectManagerTitle_recommendedIdList}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_email}" for="email" />
-                        <p:inputText id="email" value="#{projectManagerController.selected.email}" title="#{bundle.CreateProjectManagerTitle_email}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_username}" for="username" />
-                        <p:inputText id="username" value="#{projectManagerController.selected.username}" title="#{bundle.CreateProjectManagerTitle_username}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_firstName}" for="firstName" />
-                        <p:inputText id="firstName" value="#{projectManagerController.selected.firstName}" title="#{bundle.CreateProjectManagerTitle_firstName}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_lastName}" for="lastName" />
-                        <p:inputText id="lastName" value="#{projectManagerController.selected.lastName}" title="#{bundle.CreateProjectManagerTitle_lastName}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_id}" for="id" />
-                        <p:inputText id="id" value="#{projectManagerController.selected.id}" title="#{bundle.CreateProjectManagerTitle_id}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_salt}" for="salt" />
-                        <p:inputText id="salt" value="#{projectManagerController.selected.salt}" title="#{bundle.CreateProjectManagerTitle_salt}" />
-                        <p:outputLabel value="#{bundle.CreateProjectManagerLabel_password}" for="password" />
-                        <p:inputText id="password" value="#{projectManagerController.selected.password}" title="#{bundle.CreateProjectManagerTitle_password}" />
-                    </p:panelGrid>
-                    <p:commandButton actionListener="#{projectManagerController.create}" value="#{bundle.Save}" update="display,:ProjectManagerListForm:datalist,:growl" oncomplete="handleSubmit(args,'ProjectManagerCreateDialog');"/>
-                    <p:commandButton value="#{bundle.Cancel}" onclick="ProjectManagerCreateDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
