--- piJEE/piJEE-web/src/main/webapp/views/back/projectManager/View.xhtml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:f="http://xmlns.jcp.org/jsf/core"
-      xmlns:p="http://primefaces.org/ui">
-
-    <ui:composition>
-
-        <p:dialog id="ProjectManagerViewDlg" widgetVar="ProjectManagerViewDialog" modal="true" resizable="false" appendTo="@(body)" header="#{bundle.ViewProjectManagerTitle}">
-            <h:form id="ProjectManagerViewForm">
-                <h:panelGroup id="display">
-                    <p:panelGrid columns="2" rendered="#{projectManagerController.selected != null}">
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_companyProfile}"/>
-                        <h:outputText value="#{projectManagerController.selected.companyProfile}" title="#{bundle.ViewProjectManagerTitle_companyProfile}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_skills}"/>
-                        <h:outputText value="#{projectManagerController.selected.skills}" title="#{bundle.ViewProjectManagerTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_skills}"/>
-                        <h:outputText value="#{projectManagerController.selected.skills}" title="#{bundle.ViewProjectManagerTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_company}"/>
-                        <h:outputText value="#{projectManagerController.selected.company}" title="#{bundle.ViewProjectManagerTitle_company}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_manager}"/>
-                        <h:outputText value="#{projectManagerController.selected.manager}" title="#{bundle.ViewProjectManagerTitle_manager}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_introduction}"/>
-                        <h:outputText value="#{projectManagerController.selected.introduction}" title="#{bundle.ViewProjectManagerTitle_introduction}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_recommendations}"/>
-                        <h:outputText value="#{projectManagerController.selected.recommendations}" title="#{bundle.ViewProjectManagerTitle_recommendations}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_activities}"/>
-                        <h:outputText value="#{projectManagerController.selected.activities}" title="#{bundle.ViewProjectManagerTitle_activities}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_curriculumVitaes}"/>
-                        <h:outputText value="#{projectManagerController.selected.curriculumVitaes}" title="#{bundle.ViewProjectManagerTitle_curriculumVitaes}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_skills}"/>
-                        <h:outputText value="#{projectManagerController.selected.skills}" title="#{bundle.ViewProjectManagerTitle_skills}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_visits}"/>
-                        <h:outputText value="#{projectManagerController.selected.visits}" title="#{bundle.ViewProjectManagerTitle_visits}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_recommendedIdList}"/>
-                        <h:outputText value="#{projectManagerController.selected.recommendedIdList}" title="#{bundle.ViewProjectManagerTitle_recommendedIdList}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_email}"/>
-                        <h:outputText value="#{projectManagerController.selected.email}" title="#{bundle.ViewProjectManagerTitle_email}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_username}"/>
-                        <h:outputText value="#{projectManagerController.selected.username}" title="#{bundle.ViewProjectManagerTitle_username}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_firstName}"/>
-                        <h:outputText value="#{projectManagerController.selected.firstName}" title="#{bundle.ViewProjectManagerTitle_firstName}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_lastName}"/>
-                        <h:outputText value="#{projectManagerController.selected.lastName}" title="#{bundle.ViewProjectManagerTitle_lastName}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_id}"/>
-                        <h:outputText value="#{projectManagerController.selected.id}" title="#{bundle.ViewProjectManagerTitle_id}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_salt}"/>
-                        <h:outputText value="#{projectManagerController.selected.salt}" title="#{bundle.ViewProjectManagerTitle_salt}"/>
-                        <h:outputText value="#{bundle.ViewProjectManagerLabel_password}"/>
-                        <h:outputText value="#{projectManagerController.selected.password}" title="#{bundle.ViewProjectManagerTitle_password}"/>
-                    </p:panelGrid>
-                    <p:commandButton value="#{bundle.Close}" onclick="ProjectManagerViewDialog.hide()"/>
-                </h:panelGroup>
-            </h:form>
-        </p:dialog>
-
-    </ui:composition>
-</html>
