+++ piJEE/piJEE-web/src/main/webapp/views/back/question/Create.xhtml
@@ -19,7 +19,11 @@
                         <p:outputLabel value="#{bundle.CreateQuestionLabel_questionPoints}" for="questionPoints" />
                         <p:inputText id="questionPoints" value="#{questionController.selected.questionPoints}" title="#{bundle.CreateQuestionTitle_questionPoints}" />
                         <p:outputLabel value="#{bundle.CreateQuestionLabel_questionType}" for="questionType" />
-                        <p:inputText id="questionType" value="#{questionController.selected.questionType}" title="#{bundle.CreateQuestionTitle_questionType}" />
+                        <h:selectOneMenu
+                            value="#{questionController.selected.questionType}">
+                            <f:selectItem id="questionType" itemLabel="#{bundle.CreateQuestionTitle_questionType}" itemValue="" />
+                            <f:selectItems value="#{enums.questionTypes}" />
+                        </h:selectOneMenu>
                     </p:panelGrid>
                     <p:commandButton actionListener="#{questionController.create}" value="#{bundle.Save}" update="display,:QuestionListForm:datalist,:growl" oncomplete="handleSubmit(args,'QuestionCreateDialog');"/>
                     <p:commandButton value="#{bundle.Cancel}" onclick="QuestionCreateDialog.hide()"/>
