+++ piJEE/piJEE-web/src/main/webapp/views/back/quizTry/List.xhtml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+      xmlns:h="http://xmlns.jcp.org/jsf/html"
+      xmlns:f="http://xmlns.jcp.org/jsf/core"
+      xmlns:p="http://primefaces.org/ui">
+
+    <ui:composition template="/template.xhtml">
+
+
+        <ui:define name="title">
+            <h:outputText value="#{bundle.ListQuizTryTitle}"></h:outputText>
+        </ui:define>
+
+        <ui:define name="body">
+            <h:form id="QuizTryListForm">
+                <p:panel header="#{bundle.ListQuizTryTitle}">
+                    <p:dataTable id="datalist" value="#{quizTryController.items}" var="item"
+                                 selectionMode="single" selection="#{quizTryController.selected}"
+                                 paginator="true"
+                                 rowKey="#{item.idQuizTry}"
+                                 rows="10"
+                                 rowsPerPageTemplate="10,20,30,40,50"
+                                 >
+
+                        <p:ajax event="rowSelect"   update="createButton viewButton editButton deleteButton"/>
+                        <p:ajax event="rowUnselect" update="createButton viewButton editButton deleteButton"/>
+
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_idQuizTry}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.idQuizTry}"/>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_quiz}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.quiz}"/>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_percentage}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.percentage}"/>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_startDate}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.startDate}">
+                                <f:convertDateTime pattern="MM/dd/yyyy HH:mm:ss" />
+                            </h:outputText>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_finishDate}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.finishDate}">
+                                <f:convertDateTime pattern="MM/dd/yyyy HH:mm:ss" />
+                            </h:outputText>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_recording}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.recording}"/>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_candidate}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.candidate}"/>
+                        </p:column>
+                        <p:column>
+                            <f:facet name="header">
+                                <h:outputText value="#{bundle.ListQuizTryTitle_serialVersionUID}"/>
+                            </f:facet>
+                            <h:outputText value="#{item.serialVersionUID}"/>
+                        </p:column>
+                        <f:facet name="footer">
+                            <p:commandButton id="createButton" icon="ui-icon-plus"   value="#{bundle.Create}" actionListener="#{quizTryController.prepareCreate}" update=":QuizTryCreateForm" oncomplete="PF('QuizTryCreateDialog').show()"/>
+                            <p:commandButton id="viewButton"   icon="ui-icon-search" value="#{bundle.View}" update=":QuizTryViewForm" oncomplete="PF('QuizTryViewDialog').show()" disabled="#{empty quizTryController.selected}"/>
+                            <p:commandButton id="editButton"   icon="ui-icon-pencil" value="#{bundle.Edit}" update=":QuizTryEditForm" oncomplete="PF('QuizTryEditDialog').show()" disabled="#{empty quizTryController.selected}"/>
+                            <p:commandButton id="deleteButton" icon="ui-icon-trash"  value="#{bundle.Delete}" actionListener="#{quizTryController.destroy}" update=":growl,datalist" disabled="#{empty quizTryController.selected}"/>
+                        </f:facet>
+                    </p:dataTable>
+                </p:panel>
+            </h:form>
+
+            <ui:include src="Create.xhtml"/>
+            <ui:include src="Edit.xhtml"/>
+            <ui:include src="View.xhtml"/>
+        </ui:define>
+    </ui:composition>
+
+</html>
